feat: add admin productivity filter and panel components
- Implemented AdminProductivityFilter component for selecting year and month.
- Created AdminProductivityPanel component to display operational and admin productivity metrics.
- Added utility functions for handling productivity periods and year options.
- Developed tests for productivity month utilities to ensure correct functionality.

+++ supabase/migrations/00180_equipamentos_rpc_linhas_e_anos.sql
@@ -17,8 +17,8 @@ RETURNS TABLE(
   nome_loja    TEXT,
   ano          INTEGER,
   mes          INTEGER,
-  total_ordens INTEGER,
-  total_notas  INTEGER
+  total_ordens BIGINT,
+  total_notas  BIGINT
 )
 LANGUAGE plpgsql
 STABLE
